Mock j.u.Collection

+++ rv-predict-engine/src/main/java/rvpredict/db/EventInputStream.java
@@ -35,7 +35,7 @@ public class EventInputStream extends DataInputStream {
                 readLong(),
                 readInt(),
                 readLong(),
-                readLong(),
+                readInt(),
                 readLong(),
                 EventType.values()[readByte()]);
     }
